test 4b shards

+++ src/engine/engine_avx2.rs
@@ -76,8 +76,8 @@ impl Engine for Avx2 {
     }
 
     fn xor(x: &mut [u8], y: &[u8]) {
-        let x: &mut [u64] = bytemuck::cast_slice_mut(x);
-        let y: &[u64] = bytemuck::cast_slice(y);
+        let x: &mut [u32] = bytemuck::cast_slice_mut(x);
+        let y: &[u32] = bytemuck::cast_slice(y);
 
         for (x64, y64) in zip(x.iter_mut(), y.iter()) {
             *x64 ^= y64;
